Doctorant service

+++ src/main/java/com/doc/doc/service/DoctorantService.java
@@ -10,7 +10,7 @@ public interface DoctorantService {
     List<Doctorant> loadDoctorants();
 
     //Récupérez toutes les publications d'un doctorant identifié par son identifiant
-    List<Publication> loadPublications(int id);
+    List<Publication> loadPublications(Long id);
 
     //Une liste d'objets, où chaque objet contient le nombre de réunions tenues au cours
     // d'une année ainsi que le nombre d'absences du doctorant
